refactor: extract sub-components from CaptainDashboard
Extract ActiveOrderView and IncomingOrderView out of CaptainDashboard
to reduce its complexity and length. Also extract SAFETY_ITEMS to a
constants file to share between them.
