+++ src/components/captain/constants.ts
@@ -1,0 +1,7 @@
+export const SAFETY_ITEMS = [
+  'Wear safety gloves before handling fuel',
+  'Ensure no open flames or cigarettes nearby',
+  'Verify vehicle is turned OFF before fueling',
+  'Use anti-static grounding strap',
+  'Confirm fuel type matches vehicle requirements',
+];
